// src/app/models/enums/template-file.enum.ts
export enum TemplateFileEnum {
  ASSETS_INVENTORY = 'ASSETS_INVENTORY',
  THREAT_ANALYSIS = 'THREAT_ANALYSIS',
  NIST_FRAMEWORK_CORE = 'NIST_FRAMEWORK_CORE',
  IMPLEMENTATION_TIERS = 'IMPLEMENTATION_TIERS',
  PROFILE = 'PROFILE',
  RISK_ASSESSMENT = 'RISK_ASSESSMENT',
  ACTIONS_PRIORITY = 'ACTIONS_PRIORITY',
  IMPACT_RATES_PRIORITY_CODES = 'IMPACT_RATES_PRIORITY_CODES'
}