// src/app/models/enums/uploaded-file.enum.ts
export enum UploadedFileEnum {
  ASSETS_INVENTORY = 'ASSETS_INVENTORY',
  THREAT_ANALYSIS = 'THREAT_ANALYSIS',
  TARGET_PROFILE = 'TARGET_PROFILE',
  RISK_ASSESSMENT = 'RISK_ASSESSMENT',
  CURRENT_PROFILE = 'CURRENT_PROFILE',
  ACTIONS_PRIORITY = 'ACTIONS_PRIORITY',
  IMPLEMENTATION_DOC = 'IMPLEMENTATION_DOC'
}